根据社区建议修改：1.精简注释，更改注释为中文。2.减少挂载数目，核心代码放在app下3.修改develop/docker-compose.dev.ymlip限制4.排除不必要文件5.解决newline at the end of file等
