+++ .gitignore
@@ -5,3 +5,6 @@
 .env
 docker-compose.override.yml
 node_modules
+
+.vscode
+
