+++ develop/bin/dev
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-docker compose -f docker-compose.yml -f docker-compose.dev.yml up --no-deps --detach "$@"
+docker-compose -f docker-compose.yml -f docker-compose.dev.yml up --no-deps --detach "$@"
