+++ package.json
@@ -68,6 +68,7 @@
     "services/third-party-datastore",
     "services/third-party-references",
     "services/tpdsworker",
-    "services/web"
+    "services/web",
+    "services/llm"
   ]
 }
