+++ services/llm/app/routes/keys.routes.js
@@ -2,23 +2,27 @@ import express from 'express';
 import { KeysController } from '../controllers/keys.controller.js';
 
 const router = express.Router();
-const controller = new KeysController(); // 创建控制器实例
+const controller = new KeysController();
 
-// 添加服务商
+// add or update 
 router.post('/keys', controller.saveKey.bind(controller));
 
-// 删除API密钥
+// delete api key
 router.delete('/keys', controller.deleteKey.bind(controller));
 
-// 获取用户自己的llm信息
+// get user llm info
 router.get('/keys', controller.getLlmInfo.bind(controller));
 
-//获取当前使用的llm
+// get current using llm
 router.get('/usingLlm', controller.getUsingLlm.bind(controller));
 
+// get current using llm with info
 router.put('/usingLlm', controller.updateUsingLlm.bind(controller));
 
+// update using chat or completion model for the given api key name
 router.put('/usingModel', controller.updateUsingModel.bind(controller));
 
 export default router;
+
+
