+++ services/llm/app/routes/llm.routes.js
@@ -12,3 +12,6 @@ router.post('/completion', controller.completion.bind(controller));
 
 
 export default router;
+
+
+
