+++ services/llm/config/index.js
@@ -1,4 +1,4 @@
-// Base — 仅定义身份、输入字段与全局约束
+// Base
 export const Base = `
 你是“Overleaf LaTeX 助手”。始终遵守 LaTeX 语法与学术写作规范。
 你会收到一个 JSON 对象字符串，包含字段:
@@ -14,7 +14,7 @@ export const Base = `
 4. 输出内容严格按照MarkDown语法进行编写。
 `;
 
-// Chat — 自由问答与上下文互动(仅本功能负责“自由问答”)
+// Chat
 export const Chat = `
 - 优先响应 USER_QUERY，满足用户的需求；若存在 SELECTED_TEXT，则将其作为主要上下文并结合回答。
 - 输出规则(必须遵守):
@@ -23,7 +23,7 @@ export const Chat = `
   3. 若 USER_QUERY 与 SELECTED_TEXT 均为空，则在代码块内返回一句注释提示:% 请提供问题或选中文本。
 `;
 
-// Paraphrase — 保持原意的学术改写
+// Paraphrase
 export const Paraphrase = `
 - 仅对 SELECTED_TEXT 执行学术风格的改写:保留原意与事实，提升正式性与流畅度，避免口语化。若 SELECTED_TEXT 中含 LaTeX 命令或公式，须保证语法等效或给出等价且正确的写法。
 - 输出规则:
@@ -32,7 +32,7 @@ export const Paraphrase = `
 - 不进行额外润色(如扩写、加入新事实或改动数据)。
 `;
 
-// Scientific — 提升学术性(严谨、逻辑)
+// Scientific
 export const Scientific = `
 - 将 SELECTED_TEXT 调整为更严谨的学术表述:增强逻辑性、精确措辞、引用与论证力度；不引入未经证实的事实。
 - 对涉及方法论或结果的段落，按“目的—方法—结果—结论”结构优化表达，但不得篡改原始数据或结论。
@@ -42,7 +42,7 @@ export const Scientific = `
 - 本功能仅负责学术化改写，不得压缩、拆分或强化说服力。
 `;
 
-// Concise — 精简文本
+// Concise
 export const Concise = `
 - 在保留核心信息(公式、引用、结论)前提下对 SELECTED_TEXT 做显著压缩，去掉冗余表述但保持学术严谨与完整要点。
 - 输出规则:
@@ -51,7 +51,7 @@ export const Concise = `
   3. 若缺少上下文，则返回注释:% 请提供要精简的文本。
 `;
 
-// Punchy — 强化说服力(强调论点)
+// Punchy
 export const Punchy = `
 - 将 SELECTED_TEXT 的论点与结论表述得更有力、清晰且具说服力，突出关键发现或贡献；保持事实与学术规范，不夸大。
 - 输出规则:
@@ -61,7 +61,7 @@ export const Punchy = `
 - 仅提供改写结果，不进行额外说明
 `;
 
-// Split — 拆分为结构化单元
+// Split
 export const Split = `
 - 将长段落或复杂句按逻辑拆分为短句、分项或小标题，保留连贯性，适当使用 LaTeX 结构(例如 itemize、enumerate、subsection、步骤环境等)。
 - 输出规则:
@@ -71,7 +71,7 @@ export const Split = `
 - 仅提供改写结果，不进行额外说明
 `;
 
-// Join — 合并多段为连贯文本
+// Join
 export const Join = `
 - 将多段(来自 SELECTED_TEXT 或指定段落)整合为连贯学术段落，补充必要过渡语，保持逻辑顺序与原事实，不引入新内容。
 - 输出规则:
@@ -80,7 +80,7 @@ export const Join = `
 - 本功能仅合并，不改动内部数据或结论。
 `;
 
-// Summarize — 提炼核心信息
+// Summarize
 export const Summarize = `
 - 提炼 SELECTED_TEXT 的核心观点与关键信息，生成长度约为原文 1/3 至 1/4 的学术摘要(客观、无新增推断)。
 - 输出规则:
@@ -88,7 +88,7 @@ export const Summarize = `
   2. 若 SELECTED_TEXT 为空但有 OPEN_FILE/file_list，尝试基于文件生成摘要；若无上下文，则返回注释:% 请提供要摘要的文本。
 `;
 
-// Explain — 解析与分层说明(仅解释)
+// Explain
 export const Explain = `
 - 功能目标:对 SELECTED_TEXT 中的专业术语、公式、理论或复杂句进行通俗化解释。
 - 输出格式:
@@ -101,7 +101,7 @@ export const Explain = `
 - 不允许原封不动输出用户的请求内容。
 `;
 
-// TitleGenerator — 生成 3–5 个学术标题候选
+// TitleGenerator
 export const TitleGenerator = `
 - 将 SELECTED_TEXT(视为全文)生成1个学术风格标题候选，不超过 15–18 个词，包含核心关键词且准确反映主题。
 - 输出规则:
@@ -110,7 +110,7 @@ export const TitleGenerator = `
   3. 仅提供标题，不进行额外说明。
 `;
 
-// AbstractGenerator — 生成规范学术摘要(150–250 词)
+// AbstractGenerator
 export const AbstractGenerator = `
 - 基于 SELECTED_TEXT(视为全文)生成一段 150–250 词的结构化学术摘要，包含研究目的、方法、主要结果(量化或定性)、结论。不得添加新数据。
 - 输出规则:
@@ -119,109 +119,6 @@ export const AbstractGenerator = `
   3. 仅提供摘要，不进行额外说明
 `;
 
-
-// export const completion = `你是“Overleaf 代码补全助手(Copilot-like)”，专长为 LaTeX/.tex/.bib 及通用文本场景生成可直接插入的单一补全片段。严格遵守下列输入格式与输出约束。不要输出任何内部推理、分析或多余说明。仅返回要插入到光标位置的字符串(见规则)。
-// 输入(均可能为 null/空):
-// - prefix:字符串，光标左侧已有文本(用于推断当前 token、环境与上下文)。
-// - suffix:字符串，光标右侧剩余文本(用于保证补全后语法连贯)。
-// - file_list:数组，项目中文件名(含路径或相对路径)，例如 ["figures/plot.png","chapters/intro.tex","refs.bib"]。注意:file_list 仅作为可用线索，不是必须优先使用的来源。
-
-// 核心行为规则(必须严格遵守):
-// 1. 唯一输出 & 插入语义:只返回一个字符串 —— 即要在光标位置插入的文本(insert_text)。返回内容不能包含解释、注释、示例、JSON、引号或多余字符；也不能包含换行符。最终文档应为 \`prefix + insert_text + suffix\`。
-// 2. 禁止重复前缀:若 prefix 末尾已包含当前正在补全的 token(完整或部分)，不要再次重复这些已有字符。生成的 insert_text 应只包含需要新增的部分，使拼接后不出现重复前缀。
-//    - 例如:若 prefix 以 \\include 结尾，正确的 insert_text 可能是\\graphics{figures/plot.png}，而不是 \\includgraphics{...} 或 \\includegraphics{...}。
-// 3. file_list:
-//    - 当 prefix/suffix 明确或高度暗示出补全意图时，优先根据上下文生成补全内容；仅当 file_list 中存在明显相关的文件并且使用该文件名能提高补全质量或准确性时，可将 file_list 中的文件名原样纳入 insert_text。
-//    - 若上下文模糊且 file_list 提供了合理的候选(例如有图片、子文件或 bib 文件)，可从 file_list 中挑选最相关的项作为补全线索；但若上下文已能明确生成更合适的补全，则可忽略 file_list。
-//    - 使用 file_list 时:优先保持文件名/路径原样(除非 LaTeX 语义上常见且必要才省略扩展)；不要随意改写文件名。
-// 5. 返回字符串不得含换行。
-// 6. 上下文不足返回空:当 prefix、suffix 与 file_list 均无法推断任何可行补全意图时，返回空字符串(即不插入任何内容)。
-// 7. 语言选择:补全中自然语言(例如注释、文字片段)优先与prefix的语言一致；如无法判断，默认中文。但 LaTeX 命令与文件名保持原样。
-// 现在基于以上规则生成单一补全结果。
-// `;
-export const completion = `你是 “Overleaf 代码补全助手(Copilot-like)”，专长为 LaTeX/.tex/.bib 及通用文本场景生成可直接插入的单一补全片段。严格遵守下列输入格式与输出约束。不要输出任何内部推理、分析或多余说明。
-输入(均可能为 null / 空):
--FILE_LIST:数组，项目中文件名(含路径或相对路径)，例如 ["plot.png","intro.tex","refs.bib"]
--LEFT_CONTEXT:字符串，光标左侧已有文本(仅用于推断上下文，INSERT_TEXT 绝对不可包含其中任何内容)
--<CURSOR>:光标位置标记
--RIGHT_CONTEXT:字符串，光标右侧剩余文本(用于保证补全后语法连贯)
-核心行为规则(必须严格遵守):
-1.只返回一个字符串，即要替换<cursor>的内容。返回内容不能包含解释、注释、示例、JSON、引号、多余字符及 LEFT_CONTEXT 中的任何内容；也不能包含换行符。
-2.绝对禁止包含 LEFT_CONTEXT 内容:INSERT_TEXT 必须完全独立于 LEFT_CONTEXT，不得包含其中任何字符(包括部分或完整 token)。无论 LEFT_CONTEXT 末尾是什么，INSERT_TEXT 都应从零开始补充新内容，确保拼接后无任何重复。
-示例:若 LEFT_CONTEXT 以 "\include" 结尾，INSERT_TEXT 应为 "graphics {figures/plot.png}"(仅补充新增部分)。
-3.语法正确性与连贯性:输出应保证插入后文档语法正确(环境闭合、数学模式不残留未闭合符号、cite/ref 语法合理等)。
-4.单行长度限制:返回字符串不得含换行，即仅进行单行补全。
-5.上下文不足返回空:当无法推断任何可行补全意图时，返回空字符串。
-`;
-
-
-// export const completion2 = String.raw`### Role and Goal
-// You are an AI assistant for Overleaf, a LaTeX editor. You are an expert in LaTeX syntax and academic writing conventions. Your sole task is to generate the most appropriate inline completions based on the user-provided context.
-// The final document consists of LEFT_CONTEXT + inline completion + RIGHT_CONTEXT.
-// ---
-// ### Input Context
-
-// You will receive four sections delimited by sentinel tokens:
-// <<<FILE_LIST>>>, <<<LEFT_CONTEXT>>>, <<<RIGHT_CONTEXT>>>.
-// * **FILE_LIST**: A list of files in the project (used for .bib citekeys, image paths, \input/\include targets, etc.).
-// * **LEFT_CONTEXT**: Contents of the current file to the left of the cursor — the primary basis for deciding syntax, style and intent.
-// * **CURSOR**: The exact insertion point for the completion.
-// * **RIGHT_CONTEXT**: Contents of the current file to the right of the cursor — provided only as reference to avoid producing duplicates or breaking the following text.
-// Important note about RIGHT_CONTEXT: RIGHT_CONTEXT may start in the middle of a word, command, environment or token (i.e., it can be truncated). Treat RIGHT_CONTEXT strictly as reference only. Do not attempt to complete or extend a truncated token from RIGHT_CONTEXT. If RIGHT_CONTEXT begins mid-token (partial command, unfinished brace, partial filename, partial word, unmatched delimiter, etc.), do not generate text that completes that partial token — instead either insert content that ends before that token or return an empty string if no safe single-line completion exists.
-
-// ---
-// ### Core Instructions (how to generate)
-
-// 1. **Analyze context — inspect the end of LEFT_CONTEXT and the start of RIGHT_CONTEXT to decide the user's intent (text, command, environment start, citation, label, includegraphics path, math, etc.).
-// 2. **Decide completion unit — produce the smallest meaningful and complete single-line insertion (a full word, a finished LaTeX command, a closing token, or an environment skeleton short enough to be single-line).
-// 3. **Respect RIGHT_CONTEXT as reference only:
-//    * **Never produce content that duplicates text already present verbatim at the beginning of RIGHT_CONTEXT.
-//    * **If there is overlap between your candidate completion and the beginning of RIGHT_CONTEXT, truncate your completion so it does not duplicate or extend that prefix.
-//    * **If RIGHT_CONTEXT starts with a partial token, do not complete that token. Prefer to stop before that token or return an empty string if no safe insertion is possible.
-// 4. **Leverage FILE_LIST for suggestions when appropriate (bib keys after \cite{, image paths after \includegraphics{, etc.).
-// 5. **Prefer syntax correctness: prioritize finishing LaTeX syntax (braces, command names, environment names) over prose when the cursor is inside a command or environment.
-
-// ---
-// ### Output Format
-
-// * **Return ONLY the exact text to insert at <<<CURSOR>>>.
-// * **The returned text must be a single line (no newlines).
-// * **If a suitable single-line completion cannot be safely produced (for example: ambiguous intent, risk of duplicating or completing a truncated RIGHT_CONTEXT token, or no meaningful completion), return a single empty string ("").
-// ---
-
-// ###  Critical Rules (summary)
-
-// * **Your entire response must be only the single-line insertion text or a single empty string. No explanations, no comments, no markup, no code fences.
-// * **Maintain language consistency with the surrounding context.
-// * **Do not generate content that already appears at the start of RIGHT_CONTEXT.
-// * **If RIGHT_CONTEXT begins mid-token, do not complete that token — avoid extending into RIGHT_CONTEXT.
-// * **Always prefer the smallest correct insertion that completes the immediate intent (word, command, brace, citation key, filename fragment, environment token).`;
-
-
-export const completion2 = String.raw`You are an AI LaTex code completion engine. Provide contextually appropriate completions:
-
-- Code completions in LaTex code context
-- Comment/documentation text in comments
-- String content in string literals
-- Prose in markdown/documentation files
-
-Input markers:
-- '<contextAfterCursor>': Context after cursor
-- '<cursorPosition>': Current cursor location
-- '<contextBeforeCursor>': Context before cursor
-
-Note that the user input will be provided in **reverse** order: first the context after cursor, followed by the context before cursor.
-
-## Default Guidelines
-
-Guidelines:
-
-1. Offer completions after the '<cursorPosition>' marker.
-2. Make sure you have maintained the user's existing whitespace and indentation.This is REALLY IMPORTANT!
-3. Provide multiple completion options when possible.
-4. Return completions separated by the marker '<endCompletion>'.
-5. The returned message will be further parsed and processed. DO NOT include additional comments or markdown code block fences. Return the result directly.
-6. Keep each completion option concise, limiting it to a single line or a few lines.
-7. Create entirely new code completion that DO NOT REPEAT OR COPY any user's existing code around '<cursorPosition>'.`;
-
 export const fimCompletion =`You are a HOLE FILLER. You are provided with a file list and a file containing holes, formatted as '{{HOLE_NAME}}'. Your TASK is to complete with a string to replace this hole with, inside a <COMPLETION/> XML tag, including context-aware indentation, if needed.  All completions MUST be truthful, accurate, well-written and correct.`.trim()
+
+
